requirements.txt
appdirs==1.4.3
appnope==0.1.0
astroid==1.5.2
attrs==16.0.0
backports-abc==0.5
backports.functools-lru-cache==1.3
backports.shutil-get-terminal-size==1.0.0
bleach==1.5.0
cachetools==2.0.0
certifi==2017.1.23
cffi==1.7.0
click==6.7
configparser==3.5.0
cryptography==1.4
cssselect==1.0.1
cycler==0.10.0
decorator==4.0.11
dill==0.2.6
dj-database-url==0.3.0
Django==1.8.1
django-postgrespool==0.3.0
entrypoints==0.2.2
enum34==1.1.6
Flask==0.12
functools32==3.2.3.post2
future==0.16.0
futures==3.1.1
gapic-google-cloud-datastore-v1==0.15.3
gax-google-logging-v2==0.8.3
gax-google-pubsub-v1==0.8.3
gcloud==0.17.0
google-api-python-client==1.5.0
google-auth==1.0.1
google-auth-httplib2==0.0.2
google-cloud-core==0.24.1
google-cloud-datastore==1.0.0
google-gax==0.15.12
googleapis-common-protos==1.5.2
grpc-google-logging-v2==0.8.1
grpc-google-pubsub-v1==0.8.1
grpcio==1.3.0
gunicorn==19.3.0
html2text==2016.9.19
html5lib==0.9999999
httplib2==0.10.3
idna==2.1
image==1.5.5
ipaddress==1.0.16
ipykernel==4.5.2
ipython==5.3.0
ipython-genutils==0.1.0
ipywidgets==6.0.0
isort==4.2.5
itsdangerous==0.24
Jinja2==2.9.5
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.0.0
jupyter-console==5.1.0
jupyter-core==4.3.0
jws==0.1.3
lazy-object-proxy==1.3.1
lxml==3.7.3
MarkupSafe==0.23
matplotlib==2.0.0
mccabe==0.6.1
mistune==0.7.3
nbconvert==5.1.1
nbformat==4.3.0
notebook==4.4.1
numpy==1.12.0
oauth2client==3.0.0
olefile==0.44
packaging==16.8
pandas==0.19.2
pandocfilters==1.4.1
parsel==1.1.0
pathlib2==2.2.1
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.1.0
ply==3.8
prompt-toolkit==1.0.13
proto-google-cloud-datastore-v1==0.90.4
protobuf==3.3.0
psycopg2==2.6
ptyprocess==0.5.1
pyasn1==0.2.3
pyasn1-modules==0.0.8
pycparser==2.14
pycryptodome==3.4.3
PyDispatcher==2.0.5
Pygments==2.2.0
pylint==1.7.1
pymongo==3.4.0
pyOpenSSL==16.0.0
pyparsing==2.2.0
Pyrebase==3.0.27
python-dateutil==2.6.0
python-jwt==2.0.1
pytz==2016.10
pyzmq==16.0.2
qtconsole==4.2.1
queuelib==1.4.2
requests==2.14.2
requests-toolbelt==0.7.0
rsa==3.4.2
scandir==1.5
scikit-learn==0.18.1
scipy==0.18.1
Scrapy==1.3.3
service-identity==16.0.0
simplegeneric==0.8.1
simplejson==3.8.2
singledispatch==3.4.0.3
six==1.10.0
sklearn==0.0
SQLAlchemy==1.0.4
subprocess32==3.2.7
terminado==0.6
testpath==0.3
tornado==4.4.2
traitlets==4.3.2
Twisted==16.3.2
uritemplate==0.6
virtualenv==15.1.0
w3lib==1.15.0
wcwidth==0.1.7
Werkzeug==0.11.15
whitenoise==1.0.6
widgetsnbextension==2.0.0
wrapt==1.10.10
zope.interface==4.2.0
